Make creating of kernel backups optional
This change makes creation of kernel backups optional, using a new option in the default config file 'CREATE_BACKUP'. It defaults to "true", so existing functionality (always creating backups) does not change for users. This is helpful for folks, like myself, that have limited /boot size and have found ourselves unable to boot the system because sbupdate failed to create a new image due to not having enough free space on the partition to write a new image.